serializer refactor
Work in progress branch, that will replace the master branch in the futur.
The goal of this new branch is to serialize datas into ScritableObjects instead of directly build prefabs. By doing this, its easier to debug and reverse engineer. This refactor also come with a huge code clean up and reorganisation, so the code is easier to understand and modify.